refactored send interaction response in specific module
