+++ lib/vox/command.ex
@@ -29,34 +29,4 @@ defmodule Vox.Command do
           Nostrum.Struct.ApplicationCommandInteractionData.options() | nil
   def get_option(interaction, name),
     do: Enum.find(interaction.data.options || [], fn %{name: n} -> n == name end)
-
-  @spec send_ephemeral(Interaction.t(), binary()) :: {:ok}
-  def send_ephemeral(%Interaction{} = interaction, content) do
-    Logger.debug("Sending ephemeral response")
-
-    response = %{
-      type: 4,
-      data: %{
-        content: content,
-        # For ephemeral message
-        flags: 1 <<< 6
-      }
-    }
-
-    Api.create_interaction_response(interaction, response)
-  end
-
-  @spec send_response(Interaction.t(), binary()) :: {:ok}
-  def send_response(%Interaction{} = interaction, content) do
-    Logger.debug("Sending response")
-
-    response = %{
-      type: 4,
-      data: %{
-        content: content
-      }
-    }
-
-    Api.create_interaction_response(interaction, response)
-  end
 end
